backend/requirements.txt
uvicorn>=0.23.2, <0.24
fastapi>=0.101, <1.0
requests>=2.31.0, <2.32
httpx>=0.24.1, <0.25
asgi-correlation-id>=4.2.0, <4.3
python-decouple>=3.8, <3.9
redis>=4.5.5, <4.6
pytest>=7.4.0, <7.5
python-dateutil>=2.8.2, <2.9